+++ Packages/dev.unity2foxglove.sdk/Runtime/Schemas/Proto/Publishers/FoxglovePointCloudPublisher.cs
@@ -8,7 +8,6 @@ using System;
 using System.Collections.Generic;
 using System.Threading;
 using Foxglove.Schemas;
-using Foxglove.Schemas.PointCloud;
 using UnityEngine;
 using Unity.FoxgloveSDK.Schemas;
 using Unity.FoxgloveSDK.Schemas.PointCloud;